Adding some calculations and testing some new features.

+++ Application/OBD2-Display/TestData.cs
@@ -5,7 +5,7 @@ namespace OBD2_Display
     public class TestData
     {
         [JsonProperty("time")]
-        public long Time { get; set; }
+        public float Time { get; set; }
 
         [JsonProperty("rpm")]
         public int RPM { get; set; }
